Reorganize dataCollectors structure
