+++ src/main/services/MiningModule.groovy
@@ -8,8 +8,11 @@ import interfaces.OutputProcessor
 import interfaces.ProjectProcessor
 
 import services.commitFilters.InCommitListAndHasMutuallyModifiedMethodsFilter
+import services.dataCollectors.BuildRequester
+import services.dataCollectors.MergeConflictCollector
+import services.dataCollectors.StatisticsCollector
+import services.dataCollectors.modifiedLinesCollector.ModifiedLinesCollector
 
-import services.modifiedLinesCollector.ModifiedLinesCollector
 import services.projectProcessors.FilterNonExistentProjectsProcessor
 import services.projectProcessors.ForkAndEnableTravisProcessor
 
@@ -22,7 +25,7 @@ public class MiningModule extends AbstractModule {
 
         dataCollectorBinder.addBinding().to(MergeConflictCollector.class)
         dataCollectorBinder.addBinding().to(ModifiedLinesCollector.class)
-        dataCollectorBinder.addBinding().to(StatisticsCollectorImpl.class)
+        dataCollectorBinder.addBinding().to(StatisticsCollector.class)
         dataCollectorBinder.addBinding().to(BuildRequester.class)
 
         bind(CommitFilter.class).to(InCommitListAndHasMutuallyModifiedMethodsFilter.class)
