+++ src/test/integration/TestModule.groovy
@@ -5,11 +5,9 @@ import interfaces.CommitFilter
 import interfaces.ProjectProcessor
 import interfaces.OutputProcessor
 import services.commitFilters.InCommitListAndHasMutuallyModifiedMethodsFilter
-
+import services.dataCollectors.StatisticsCollector
 import util.*
-import services.*
-
-import services.modifiedLinesCollector.ModifiedLinesCollector
+import services.dataCollectors.modifiedLinesCollector.ModifiedLinesCollector
 
 import com.google.inject.*
 import com.google.inject.multibindings.Multibinder
@@ -22,7 +20,7 @@ public class TestModule extends AbstractModule {
         Multibinder<ProjectProcessor> projectProcessorBinder = Multibinder.newSetBinder(binder(), ProjectProcessor.class)
 
         dataCollectorBinder.addBinding().to(ModifiedLinesCollector.class);
-        dataCollectorBinder.addBinding().to(StatisticsCollectorImpl.class);
+        dataCollectorBinder.addBinding().to(StatisticsCollector.class);
         bind(CommitFilter.class).to(InCommitListAndHasMutuallyModifiedMethodsFilter.class)
 
         projectProcessorBinder.addBinding().to(EmptyProjectProcessor.class)
